+++ pythonFiles/include/blender_vscode/communication.py
@@ -9,7 +9,7 @@ import debugpy
 import flask
 import requests
 
-from .environment import LOG_FLASK, blender_path, scripts_folder, python_path
+from .environment import LOG_FLASK, VSCODE_IDENTIFIER, blender_path, scripts_folder, python_path
 from .utils import run_in_main_thread
 from . import log
 
@@ -132,6 +132,7 @@ def send_connection_information(path_mappings: Dict):
             "blenderPath": str(blender_path),
             "scriptsFolder": str(scripts_folder),
             "addonPathMappings": path_mappings,
+            "vscodeIdentifier": VSCODE_IDENTIFIER,
         }
     )